pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cffi==2.0.0b1",
    "dj-database-url>=3.0.1",
    "django-allauth[headless-spec,mfa]>=64.0",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "django-ninja-extra>=0.30.1",
    "django-ninja-jwt>=5.3.7",
    "django-otp>=1.6.1",
    "django-phonenumber-field[phonenumberslite]>=8.1.0",
    "django-two-factor-auth>=1.17.0",
    "pydantic==2.12.0a1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "typing-extensions==4.15.0rc1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "ruff>=0.12.10",
]


[tool.black]
line-length = 79
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    .git
  | __pycache__
  | docs/source/conf.py
  | old
  | build
  | dist
  | .venv
  | frontend
)/
'''
